Keep module mocks sorted, record mock values
This changes the format of expoModules to be formatted in a way that is a little more natural to read and work with. It also sorts module names and their properties so that changes to the native side don't result in big diffs.

This also adds support for recording mock values so that the Jest tests can use values from a real phone. Right now most fields are sanitized out since many had maybe-sensitive data like device IDs and file paths.

The mock printing code now runs in componentDidMount in the app so that if there's an error early on, we get a better redbox about it.

Test plan: tested with new-project-template

fbshipit-source-id: faa6622