Remove Modal swap (replacing Expo.Modal with RN.Modal) from jest-expo
In jest-expo we replaced Expo.Modal with RN.Modal, which I believe is unnecessary. Expo.Modal isn't documented and we don't override RN.Modal with it (and if we did, we'd want to override RN.Modal with Expo.Modal, the other way around).

We need to remove this because Jest runs this setup before it runs each test file (including setting up the mocks). So if a test file tries to mock something, this setup file runs before it and messes up our module state (in this case, some of the logging modules).

Test Plan: Ran expo-sdk tests (with new logging tests, coming in a future commit) with this change, tests pass now.

fbshipit-source-id: 445298b